Phase 2: fix confirmed correctness defects (D1-D2, D5-D7)
Implements Phase 2 of RELEASE_PLAN_V3.md. Six defects, two of them severe.

D1 - Non-contiguous tensors were silently misread (T2.1, T2.2)

fit_t/transform_t/fit_transform_t validated rank, dtype and element count,
then walked data_ptr() + numel(). A 1-D non-contiguous view has a stride
greater than one, so this read unrelated memory and produced wrong cut
points with no error raised. A column view of a 2-D dataset - the most
natural way to feed one feature at a time - hit this exactly.

Resolution: accept non-contiguous input via contiguous() rather than
reject it. Rejecting would force every caller to add .contiguous() by
hand; contiguous() is a no-op when the input already is, so the common
path costs nothing. Non-CPU tensors are now rejected instead of having
their data_ptr() dereferenced.

The three duplicated validation blocks are consolidated into
validate_tensor() and validate_pair(). The singular/plural empty-tensor
messages are preserved deliberately - they are correct for their
respective arities. Every pre-existing exception message is unchanged, so
no existing test needed editing.

D2 - Discretizer::direction was never initialized (T2.3)

Only the parameterized CPPFImdlp constructor and BinDisc::fit assigned it,
so CPPFImdlp() = default left it indeterminate and transform() branched on
an uninitialized value. Now defaults to bound_dir_t::RIGHT. Note that
bound_dir_t::LEFT is still never assigned anywhere, leaving the
lower_bound branch unreachable; building that out is T6.1.

D5 - Cache keys narrowed size_t to int (T2.6)
D6 - entropy() size guard wrapped around on inverted intervals (T2.4)
D7 - Signed/unsigned comparison in BinDisc::linspace (T2.5)

Verification

85/85 tests pass: the 82 pre-existing tests are unmodified, plus 3 new
regression tests. Debug and release builds are warning-free. Coverage of
src/ is 99.1% lines and 100% functions; the four uncovered lines are gcov
closing-brace artifacts in code this commit does not touch.

The two contiguity tests were confirmed to fail with contiguous()
temporarily reverted. The direction test cannot be shown to fail that way:
the pre-fix behaviour was indeterminate rather than reliably wrong, so it
passed before the fix too. It locks in the now-defined behaviour rather
than demonstrating the old bug.

+++ src/BinDisc.cpp
@@ -67,7 +67,7 @@ namespace mdlp {
         }
         precision_t delta = (end - start) / static_cast<precision_t>(num - 1);
         std::vector<precision_t> linspc;
-        for (size_t i = 0; i < num; ++i) {
+        for (size_t i = 0; i < static_cast<size_t>(num); ++i) {
             precision_t val = start + delta * static_cast<precision_t>(i);
             linspc.push_back(val);
         }
